requirements.txt
azure-functions
requests
openai